change data type of contribution count from int to long (#219)
## Checklist
- [ ] 충분한 양의 자동화 테스트를 작성했는가?
  - 계단정복지도 서비스는 사이드 프로젝트로 진행되는 만큼 충분한 QA 없이 배포되는 경우가 많습니다. 따라서 자동화 테스트를 꼼꼼하게 작성하는 것이 서비스 품질을 유지하는 데 매우 중요합니다.

+++ app-server/subprojects/bounded_context/challenge/application/src/main/kotlin/club/staircrusher/challenge/application/port/in/use_case/GetCountForNextChallengeRankUseCase.kt
@@ -20,7 +20,7 @@ class GetCountForNextChallengeRankUseCase(
     private val challengeRankRepository: ChallengeRankRepository,
     private val challengeService: ChallengeService,
 ) {
-    fun handle(challengeId: String, userId: String): Int?  = transactionManager.doInTransaction {
+    fun handle(challengeId: String, userId: String): Long?  = transactionManager.doInTransaction {
         userApplicationService.getUser(userId) ?: throw SccDomainException("잘못된 계정입니다.")
         val challenge = challengeRepository.findByIdOrNull(challengeId) ?: throw SccDomainException("잘못된 챌린지입니다.")
         if (!challengeService.hasJoined(userId, challengeId)) {
